refactoring the spacy name title detector
