+++ scrubadub/detectors/spacy.py
@@ -5,10 +5,11 @@ import logging
 import importlib
 
 from wasabi import msg
-from typing import Generator, Iterable, Optional, Sequence, List
+from typing import Generator, Iterable, Optional, Sequence, List, Set
 
 try:
     import spacy
+    import spacy.tokens
 except ImportError as e:
     if e.name == "spacy":
         raise ImportError(
@@ -121,6 +122,8 @@ class SpacyEntityDetector(Detector):
         # Only enable necessary pipes
         self.nlp.select_pipes(enable=["transformer", "tagger", "parser", "ner"])
 
+        self.yielded_filth = None  # type: Optional[Set[str]]
+
     @staticmethod
     def check_spacy_version() -> bool:
         """Ensure that the version od spaCy is v3."""
@@ -168,6 +171,57 @@ class SpacyEntityDetector(Detector):
             document_list[i_doc] = re.sub(whitespace_regex, ' ', text)
         return document_list
 
+    def _run_spacy(self, document_list: Sequence[str], document_names: Sequence[Optional[str]]) -> List[spacy.tokens.doc.Doc]:
+        i = 0
+        spacy_docs = []  # type: List[spacy.tokens.doc.Doc]
+        generator = self.nlp.pipe(document_list)
+        while True:
+            try:
+                spacy_doc = next(generator)
+            except IndexError as e:
+                if e.args[0] == 'index out of range in self':
+                    message = "Error processing documents due to spacy's transformer model. To use this model, try " \
+                              "preprocessing the text by removing non-words and reducing spaces. Skipping file: {}"
+                    logger = logging.getLogger('scrubadub.detectors.spacy.SpacyEntityDetector')
+                    logger.warning(message.format(document_names[i]))
+                    spacy_doc = list(self.nlp.pipe(' '))[0]
+                else:
+                    raise e
+            except StopIteration:
+                break
+            i += 1
+            spacy_docs.append(spacy_doc)
+
+        return spacy_docs
+
+    def _yield_filth(self, doc_name: str, text: str, ent: spacy.tokens.span.Span) -> Generator[Filth, None, None]:
+        if ent.label_ not in self.named_entities:
+            return
+        filth_class = self.filth_cls_map.get(ent.label_, Filth)
+        if self.preprocess_text:
+            # When yielding the filth we need to yield filth as found in the original un-preprocessed text.
+            # This section searches for text with the inverse of the preprocessing step.
+            if ent.text in self.yielded_filth:
+                return
+            self.yielded_filth.add(ent.text)
+
+            class SpacyEntDetector(RegexDetector):
+                filth_cls = filth_class
+                regex = re.compile(re.escape(ent.text).replace('\\ ', r'\s+'))
+
+            regex_detector = SpacyEntDetector(name=self.name, locale=self.locale)
+            yield from regex_detector.iter_filth(text, document_name=doc_name)
+        else:
+            # If we didn't pre-process, just return the filth as it was found.
+            yield filth_class(
+                beg=ent.start_char,
+                end=ent.end_char,
+                text=ent.text,
+                document_name=(str(doc_name) if doc_name else None),  # None if no doc_name provided
+                detector_name=self.name,
+                locale=self.locale,
+            )
+
     def iter_filth_documents(self, document_list: Sequence[str],
                              document_names: Sequence[Optional[str]]) -> Generator[Filth, None, None]:
         """Yields discovered filth in a list of documents.
@@ -186,55 +240,13 @@ class SpacyEntityDetector(Detector):
         if self.preprocess_text:
             preprocessed_docs = self._preprocess_text(preprocessed_docs)
 
-        i = 0
-        spacy_docs = []
-        generator = self.nlp.pipe(preprocessed_docs)
-        while True:
-            try:
-                spacy_doc = next(generator)
-            except IndexError as e:
-                if e.args[0] == 'index out of range in self':
-                    message = "Error processing documents due to spacy's transformer model. To use this model, try " \
-                              "preprocessing the text by removing non-words and reducing spaces. Skipping file: {}"
-                    logger = logging.getLogger('scrubadub.detectors.spacy.SpacyEntityDetector')
-                    logger.warning(message.format(document_names[i]))
-                    spacy_doc = list(self.nlp.pipe(' '))[0]
-                else:
-                    raise e
-            except StopIteration:
-                break
-            i += 1
-            spacy_docs.append(spacy_doc)
+        spacy_docs = self._run_spacy(document_list=preprocessed_docs, document_names=document_names)
 
-        yielded_filth = set()
+        self.yielded_filth = set()
         for doc_name, doc, text in zip(document_names, spacy_docs, document_list):
             for ent in doc.ents:
-                if ent.label_ not in self.named_entities:
-                    continue
-                filth_class = self.filth_cls_map.get(ent.label_, Filth)
-                if self.preprocess_text:
-                    # When yielding the filth we need to yield filth as found in the original un-preprocessed text.
-                    # This section searches for text with the inverse of the preprocessing step.
-                    if ent.text in yielded_filth:
-                        continue
-                    yielded_filth.add(ent.text)
-
-                    class SpacyEntDetector(RegexDetector):
-                        filth_cls = filth_class
-                        regex = re.compile(re.escape(ent.text).replace('\\ ', r'\s+'))
-
-                    regex_detector = SpacyEntDetector(name=self.name, locale=self.locale)
-                    yield from regex_detector.iter_filth(text, document_name=doc_name)
-                else:
-                    # If we didn't pre-process, just return the filth as it was found.
-                    yield filth_class(
-                        beg=ent.start_char,
-                        end=ent.end_char,
-                        text=ent.text,
-                        document_name=(str(doc_name) if doc_name else None),  # None if no doc_name provided
-                        detector_name=self.name,
-                        locale=self.locale,
-                    )
+                yield from self._yield_filth(doc_name, text, ent)
+        self.yielded_filth = None
 
     def iter_filth(self, text: str, document_name: Optional[str] = None) -> Generator[Filth, None, None]:
         """Yields discovered filth in the provided ``text``.
